pyproject.toml
[project]
name = "file-sorter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich>=14.2.0",
    "streamlit>=1.53.0",
    "textual-fspicker>=0.6.0",
    "tk>=0.1.0",
    "typer>=0.21.1",
]
